Commit of HikariCP Connenction Test Code
