+++ boot_move/pom.xml
@@ -142,13 +142,34 @@
 			<version>2.7.6</version>
 		</dependency>
 		
-		<!-- https://mvnrepository.com/artifact/junit/junit -->
+		<!-- https://mvnrepository.com/artifact/junit/junit 
+		Junit-> @Test -->
 		<dependency>
 			<groupId>junit</groupId>
 			<artifactId>junit</artifactId>
 			<version>4.12</version>
 			<scope>test</scope>
 		</dependency>
+		
+		<!-- https://mvnrepository.com/artifact/org.springframework/spring-test 
+		spring-test -> @ContextConfiguration -->
+		<dependency>
+    		<groupId>org.springframework</groupId>
+    		<artifactId>spring-test</artifactId>
+    		<version>4.3.25.RELEASE</version>
+    		<scope>test</scope>
+		</dependency>
+
+		
+		<!-- https://mvnrepository.com/artifact/hikari-cp/hikari-cp 
+		그냥 HikariCP쓰니까 안됨 -->
+		<dependency>
+			<groupId>com.zaxxer</groupId>
+			<artifactId>HikariCP</artifactId>
+			<version>3.4.5</version>
+		</dependency>
+
+
 
         <!-- mysql이나 oracle DB 사용시 아래 설정 추가
         <dependency>
